+++ app/access-denied/page.tsx
@@ -8,8 +8,8 @@ import {
   CardDescription,
   CardHeader,
   CardTitle,
-} from '@/components/ui/card';
-import { Button } from '@/components/ui/button';
+} from '@/components/shadcn/card';
+import { Button } from '@/components/shadcn/button';
 import { ShieldX, ArrowLeft, Home, Loader2 } from 'lucide-react';
 import Link from 'next/link';
