+++ app/layout.tsx
@@ -1,7 +1,7 @@
 import './globals.css';
 import { Providers } from '@/components/providers/auth-provider';
 import { ThemeProvider } from '@/components/providers/theme-provider';
-import { Toaster } from '@/components/ui/sonner';
+import { Toaster } from '@/components/shadcn/sonner';
 
 // Auth validation handled by Keycloak provider at runtime
 // Missing/invalid env vars will produce descriptive errors when authentication is attempted
